Remove outdated comment and dependency pin on Coveralls
The version restriction for Coveralls was added in February 2020 due to a
specific issue with v1.11.0. However, Coveralls is now at v4.0.1, and the
previous concern is no longer relevant.

In the future, if we need to temporarily embargo a version, we should document
the reason in an issue and set a reminder to remove the embargo once the problem
is resolved.